backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.2
datasets==2.15.0
soundfile==0.12.1
transformers[torch]>=4.35.2
librosa==0.10.1
opencv-python==4.8.1.78
openai-whisper==20231117
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
ffmpeg-python==0.2.0
pillow==10.1.0
bcrypt==4.1.1
